pyproject.toml
[project]
name = "kiva-iccv"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pillow",
    "ipython",
    "ipykernel",
    "pip",
    "numpy",
    "pandas",
    "textwrap3",
    "importlib-metadata",
    "requests",
    "matplotlib",
    "tqdm",
    "torch",
    "torchvision",
    "pydantic"
]
